feat/rest: a little code refactor
