+++ functions/backend/routes/user.js
@@ -4,16 +4,16 @@ const { body, matchedData } = require('express-validator');
 
 const db = admin.firestore();
 
-const validateFirebaseIdToken = require('./middlewares');
-const preferences = require('./defaultPreferences');
+const preferences = require('../../util/defaultPreferences');
 
 const collection = db.collection('users');
 
 const validateUser = [body('email').isEmail().trim().escape().normalizeEmail(), body('id').trim().escape()];
 
-router.post('/create', validateFirebaseIdToken, ...validateUser, async (req, res) => {
+router.post('/create', ...validateUser, async (req, res) => {
   const sanitized = matchedData(req);
   const { email, id } = sanitized;
+
   try {
     if (!email) {
       res.status(400).send('User object is not present, or at least some properties are not present');
