+++ functions/backend/validateFirebaseToken.js
