+++ functions/util/defaultPreferences.js
